requirements.txt
datasets==2.1.0
h5py==3.8.0
hydra-core==1.3.1
numpy>=1.23.4
torch>=1.6.0
tqdm==4.63.1
transformers==4.26.0
sentencepiece==0.1.97
chardet==5.1.0
cchardet==2.1.7
matplotlib==3.7.0